FishFryToursBackend/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
future==0.18.2
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.4
Pillow==7.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.2
tzlocal==2.1
urllib3==1.25.10
